requirements.txt
python-dotenv
requests
beautifulsoup4
feedparser
google-generativeai
edge-tts
moviepy
pillow
keyring
